+++ thermohw/pymarkdown.py
@@ -44,7 +44,7 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 """
 
-from typing import TYPE_CHECKING, Tuple
+from typing import TYPE_CHECKING, Tuple, Dict
 import re
 
 from nbconvert.preprocessors import Preprocessor
@@ -61,7 +61,7 @@ class PyMarkdownPreprocessor(Preprocessor):  # type: ignore
     markdown cells with the results stored in the cell metadata.
     """
 
-    def replace_variables(self, source: str, variables: dict) -> str:
+    def replace_variables(self, source: str, variables: Dict[str, str]) -> str:
         """Replace {{variable-name}} with stored value."""
         try:
             replaced = re.sub(
@@ -72,8 +72,8 @@ class PyMarkdownPreprocessor(Preprocessor):  # type: ignore
         return replaced
 
     def preprocess_cell(
-        self, cell: "NotebookNode", resources: dict, index: int
-    ) -> Tuple["NotebookNode", dict]:
+        self, cell: "NotebookNode", resources: Dict[str, bool], index: int
+    ) -> Tuple["NotebookNode", Dict[str, bool]]:
         """Preprocess cell.
 
         Parameters